+++ ricaun.DA4R.NUnit/Revit/UI/Commands/Command.cs
@@ -1,8 +1,12 @@
+using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using ricaun.DA4R.NUnit.Revit.UI.Services;
 using ricaun.DA4R.NUnit.Services;
+using System;
+using System.IO;
+using System.Threading.Tasks;
 
 namespace ricaun.DA4R.NUnit.Revit.UI.Commands
 {
@@ -15,10 +19,27 @@ namespace ricaun.DA4R.NUnit.Revit.UI.Commands
 
             using (new CurrentDirectory())
             {
+                CopyFile(@"C:\Users\ricau\source\repos\RevitAddin.UnitTest\RevitAddin.UnitTest\bin\ReleaseFiles\RevitAddin.UnitTest 1.0.0.zip");
                 DesignAutomationController.Execute(uiapp.Application);
             }
 
             return Result.Succeeded;
+        }
+
+        private void Uiapp_Idling(object sender, Autodesk.Revit.UI.Events.IdlingEventArgs e)
+        {
+            throw new NotImplementedException();
+        }
+
+        private void CopyFile(string filePath)
+        {
+            var fileName = Path.GetFileName(filePath);
+            if (File.Exists(fileName))
+            {
+                Console.WriteLine($"Delete {fileName}");
+                File.Delete(fileName);
+            }
+            File.Copy(filePath, fileName);
         }
 
         public bool IsCommandAvailable(UIApplication applicationData, CategorySet selectedCategories)