+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] / 2022-09-19
+### Updated
+- Update ricaun.NUnit Library - Remove await/async
 ### Fixed
 - Fix UnZip File
 - Fix Test Parameters, `RevitApp` is a RevitNET type.